+++ src/context/BrowserRouterProvider/BrowserRouterProvider.tsx
@@ -4,7 +4,7 @@ import React, { useContext, useMemo } from 'react'
 import { matchPath, useHistory, useLocation, useParams } from 'react-router-dom'
 import { generateAppRoutes } from 'Routes/helpers'
 import { Route as NestedRoute } from 'Routes/helpers'
-import { routes } from 'Routes/RoutesTypes'
+import { routes } from 'Routes/RoutesCommon'
 import { usePlugins } from 'context/PluginProvider/PluginProvider'
 import { useQuery } from 'hooks/useQuery/useQuery'
