+++ src/features/defi/providers/foxy/components/FoxyManager/Overview/FoxyDetails.tsx
@@ -2,7 +2,7 @@ import { Center, Flex, ModalBody, ModalFooter, Stack, Tag } from '@chakra-ui/rea
 import { AssetNamespace, caip19 } from '@shapeshiftoss/caip'
 import { FoxyApi } from '@shapeshiftoss/investor-foxy'
 import { NetworkTypes } from '@shapeshiftoss/types'
-import { DefiParams, DefiQueryParams } from 'features/defi/contexts/DefiManagerProvider/DefiTypes'
+import { DefiParams, DefiQueryParams } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import { matchPath } from 'react-router'
 import { Amount } from 'components/Amount/Amount'
 import { AssetIcon } from 'components/AssetIcon'
