+++ src/plugins/cosmos/components/modals/GetStarted/views/GetStarted.tsx
@@ -3,7 +3,7 @@ import { Button, ModalCloseButton, VStack } from '@chakra-ui/react'
 import { CAIP19 } from '@shapeshiftoss/caip'
 import { AnimatePresence } from 'framer-motion'
 import { DefiModalHeader } from 'plugins/cosmos/components/DefiModalHeader/DefiModalHeader'
-import { StakingAction } from 'plugins/cosmos/components/modals/Staking/StakingTypes'
+import { StakingAction } from 'plugins/cosmos/components/modals/Staking/StakingCommon'
 import { useHistory, useLocation } from 'react-router-dom'
 import osmosis from 'assets/osmosis.svg'
 import { Text } from 'components/Text'
