+++ src/plugins/cosmos/components/modals/Send/views/Details.tsx
@@ -26,8 +26,8 @@ import { AccountCard } from 'components/AccountCard'
 import { Amount } from 'components/Amount/Amount'
 import { SendFormFields } from 'components/Modals/Send/Form'
 import { useSendDetails } from 'components/Modals/Send/hooks/useSendDetails/useSendDetails'
+import { SendRoutes } from 'components/Modals/Send/SendCommon'
 import { SendMaxButton } from 'components/Modals/Send/SendMaxButton/SendMaxButton'
-import { SendRoutes } from 'components/Modals/Send/SendTypes'
 import { SlideTransition } from 'components/SlideTransition'
 import { Text } from 'components/Text'
 import { TokenRow } from 'components/TokenRow/TokenRow'
