+++ src/plugins/cosmos/components/modals/Staking/StakingCommon.ts
